update common module
