+++ src/manage_voice_models.py
@@ -5,7 +5,7 @@ import zipfile
 import json
 
 from common import RVC_MODELS_DIR
-from common import display_progress
+from common import display_progress, copy_files_to_new_folder
 
 with open(
     os.path.join(RVC_MODELS_DIR, "public_models.json"), encoding="utf8"
@@ -13,23 +13,55 @@ with open(
     public_models = json.load(infile)
 
 
-def remove_suffix_after(text: str, occurrence: str):
-    location = text.rfind(occurrence)
-    if location == -1:
-        return text
+def get_current_models():
+    models_list = os.listdir(RVC_MODELS_DIR)
+    items_to_remove = ["hubert_base.pt", "MODELS.txt", "public_models.json", "rmvpe.pt"]
+    return [item for item in models_list if item not in items_to_remove]
+
+
+def load_public_models_table(predicates, progress):
+    models_table = []
+    keys = ["name", "description", "tags", "credit", "added", "url"]
+    display_progress("[~] Loading public models ...", 0.5, progress)
+    for model in public_models["voice_models"]:
+        if all([predicate(model) for predicate in predicates]):
+            models_table.append([model[key] for key in keys])
+
+    return models_table
+
+
+def load_public_model_tags():
+    return list(public_models["tags"].keys())
+
+
+def filter_public_models_table(tags, query, progress):
+
+    tags_predicate = lambda model: all(tag in model["tags"] for tag in tags)
+    query_predicate = (
+        lambda model: query.lower()
+        in f"{model['name']} {model['description']} {' '.join(model['tags'])} {model['credit']} {model['added']}".lower()
+    )
+
+    # no filter
+    if len(tags) == 0 and len(query) == 0:
+        filter_fns = []
+
+    # filter based on tags and query
+    elif len(tags) > 0 and len(query) > 0:
+        filter_fns = [tags_predicate, query_predicate]
+
+    # filter based on only tags
+    elif len(tags) > 0:
+        filter_fns = [tags_predicate]
+
+    # filter based on only query
     else:
-        return text[: location + len(occurrence)]
+        filter_fns = [query_predicate]
+
+    return load_public_models_table(filter_fns, progress)
 
 
-def copy_files_to_new_folder(file_paths, folder_path):
-    os.makedirs(folder_path)
-    for file_path in file_paths:
-        shutil.copyfile(
-            file_path, os.path.join(folder_path, os.path.basename(file_path))
-        )
-
-
-def extract_zip(extraction_folder, zip_name, remove_zip):
+def extract_model_zip(extraction_folder, zip_name, remove_zip):
     try:
         os.makedirs(extraction_folder)
         with zipfile.ZipFile(zip_name, "r") as zip_ref:
@@ -80,60 +112,6 @@ def extract_zip(extraction_folder, zip_name, remove_zip):
             os.remove(zip_name)
 
 
-def get_current_models():
-    models_list = os.listdir(RVC_MODELS_DIR)
-    items_to_remove = ["hubert_base.pt", "MODELS.txt", "public_models.json", "rmvpe.pt"]
-    return [item for item in models_list if item not in items_to_remove]
-
-
-def load_public_models_table(predicates, progress):
-    models_table = []
-    keys = ["name", "description", "tags", "credit", "added", "url"]
-    display_progress("[~] Loading public models ...", 0.5, progress)
-    for model in public_models["voice_models"]:
-        if all([predicate(model) for predicate in predicates]):
-            models_table.append([model[key] for key in keys])
-
-    return models_table
-
-
-def load_public_models_tags():
-    return list(public_models["tags"].keys())
-
-
-def load_public_models_table_w_tags(progress):
-    models_table = load_public_models_table([], progress)
-    tags = list(public_models["tags"].keys())
-    return models_table, tags
-
-
-def filter_public_models_table(tags, query, progress):
-
-    tags_predicate = lambda model: all(tag in model["tags"] for tag in tags)
-    query_predicate = (
-        lambda model: query.lower()
-        in f"{model['name']} {model['description']} {' '.join(model['tags'])} {model['credit']} {model['added']}".lower()
-    )
-
-    # no filter
-    if len(tags) == 0 and len(query) == 0:
-        filter_fns = []
-
-    # filter based on tags and query
-    elif len(tags) > 0 and len(query) > 0:
-        filter_fns = [tags_predicate, query_predicate]
-
-    # filter based on only tags
-    elif len(tags) > 0:
-        filter_fns = [tags_predicate]
-
-    # filter based on only query
-    else:
-        filter_fns = [query_predicate]
-
-    return load_public_models_table(filter_fns, progress)
-
-
 def download_online_model(url, dir_name, progress):
     if not url:
         raise Exception("Download link to model missing!")
@@ -157,7 +135,7 @@ def download_online_model(url, dir_name, progress):
 
     display_progress(f"[~] Extracting zip file...", 0.5, progress)
 
-    extract_zip(extraction_folder, zip_name, remove_zip=True)
+    extract_model_zip(extraction_folder, zip_name, remove_zip=True)
     return f"[+] Model with name '{dir_name}' successfully downloaded!"
 
 
@@ -182,7 +160,7 @@ def upload_local_model(input_paths, dir_name, progress):
         # NOTE a .pth file is actually itself a zip file
         elif zipfile.is_zipfile(input_name):
             display_progress("[~] Extracting zip file...", 0.5, progress)
-            extract_zip(output_folder, input_name, remove_zip=False)
+            extract_model_zip(output_folder, input_name, remove_zip=False)
         else:
             raise Exception(
                 "Only a .pth file or a .zip file can be uploaded by itself!"
